+++ libraries/Canvas/src/Canvas.cpp
@@ -40,14 +40,14 @@ void Canvas::PPMData(std::ostream& outputStream) const{
   }
 }
 
-inline unsigned int Canvas::convertColor(const float& ColorComponent) const{
+inline unsigned int Canvas::convertColor(const double& ColorComponent) const{
   // map the range 0.0-1.0 to 0-255 which is used for the PPM format
-  if(std::isgreater(ColorComponent, 1.0f)){return 255;}
-  else if(std::isless(ColorComponent, 0.0f)){return 0;}
+  if(std::isgreater(ColorComponent, 1.0)){return 255;}
+  else if(std::isless(ColorComponent, 0.0)){return 0;}
   else{return std::ceil(255*ColorComponent);}
 }
 
-size_t Canvas::ColorComponentToPPM(const float& ColorComponent, std::ostream& outputStream, size_t rowLineLen) const{
+size_t Canvas::ColorComponentToPPM(const double& ColorComponent, std::ostream& outputStream, size_t rowLineLen) const{
   const auto maxLineLen = 70u;
   std::string ComponentString{std::to_string(convertColor(ColorComponent))};
